py-server/requirements.txt
grpcio==1.41.0
grpcio-tools==1.41.0
PyVISA==1.11.3
protobuf==3.18.1